safari-600.7-branch/LayoutTests/fast/mediastream/MediaStreamConstructor-expected.txt

Tests webkitMediaStream.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Got local stream.
PASS localStream.getAudioTracks().length is 1
PASS localStream.getVideoTracks().length is 1
PASS typeof webkitMediaStream === 'object' is true
PASS checkIdAttribute(localStream.id) is true
PASS new webkitMediaStream(document) threw exception TypeError: Type error.
PASS new webkitMediaStream([document]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream([stream.getAudioTracks()[0], document]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream([null]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream([undefined]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream(null) threw exception TypeError: Type error.
PASS new webkitMediaStream(undefined) threw exception TypeError: Type error.

*** testing "new webkitMediaStream()"
PASS Stream constructed
PASS [object webkitMediaStream] is non-null.
PASS [object webkitMediaStream] is defined.
PASS newStream.constructor.toString() is '[object webkitMediaStreamConstructor]'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS checkIdAttribute(newStream.id) is true

*** testing "new webkitMediaStream([])"
PASS Stream constructed
PASS [object webkitMediaStream] is non-null.
PASS [object webkitMediaStream] is defined.
PASS newStream.constructor.toString() is '[object webkitMediaStreamConstructor]'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS checkIdAttribute(newStream.id) is true

*** testing "new webkitMediaStream(stream)"
PASS Stream constructed
PASS [object webkitMediaStream] is non-null.
PASS [object webkitMediaStream] is defined.
PASS newStream.constructor.toString() is '[object webkitMediaStreamConstructor]'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS checkIdAttribute(newStream.id) is true

*** testing "new webkitMediaStream([stream.getAudioTracks()[0]])"
PASS Stream constructed
PASS [object webkitMediaStream] is non-null.
PASS [object webkitMediaStream] is defined.
PASS newStream.constructor.toString() is '[object webkitMediaStreamConstructor]'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS checkIdAttribute(newStream.id) is true

*** testing "new webkitMediaStream([stream.getVideoTracks()[0]])"
PASS Stream constructed
PASS [object webkitMediaStream] is non-null.
PASS [object webkitMediaStream] is defined.
PASS newStream.constructor.toString() is '[object webkitMediaStreamConstructor]'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS checkIdAttribute(newStream.id) is true

*** testing "new webkitMediaStream([stream.getAudioTracks()[0], stream.getVideoTracks()[0]])"
PASS Stream constructed
PASS [object webkitMediaStream] is non-null.
PASS [object webkitMediaStream] is defined.
PASS newStream.constructor.toString() is '[object webkitMediaStreamConstructor]'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS checkIdAttribute(newStream.id) is true

*** testing "new webkitMediaStream([stream.getVideoTracks()[0], stream.getAudioTracks()[0], stream.getVideoTracks()[0]])"
PASS Stream constructed
PASS [object webkitMediaStream] is non-null.
PASS [object webkitMediaStream] is defined.
PASS newStream.constructor.toString() is '[object webkitMediaStreamConstructor]'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS checkIdAttribute(newStream.id) is true
PASS successfullyParsed is true

TEST COMPLETE
